Adding end date

+++ client.py
@@ -2,14 +2,21 @@ import urllib.request
 import shutil
 import os
 import time
+import sys
+import datetime
 
 URL = 'http://127.0.0.1:5000/'
 FILE_NAME = 'downloadedfile'
 DOWNLOAD_PERIOD = 100
 
 
-def run_download(url, file_name):
+def run_download(url, file_name, end_date):
     while True:
+
+        if time.time() > end_date:
+            print("Reached end date. Quitting")
+            break
+
         if os.path.exists(FILE_NAME):
             os.remove(FILE_NAME)
         else:
@@ -22,4 +29,11 @@ def run_download(url, file_name):
 
 
 if __name__ == '__main__':
-    run_download(URL, FILE_NAME)
+    year = int(sys.argv[1])
+    month = int(sys.argv[2])
+    day = int(sys.argv[3])
+
+    print("Year :" + str(year) + " Month : " + str(month) + " Day : " + str(day))
+    date = datetime.datetime(year=year, month=month, day=day).timestamp()
+
+    run_download(URL, FILE_NAME, date)
